requirements.txt
# Flask 和基礎依賴
Flask==3.0.0
Flask-CORS==4.0.0
Werkzeug>=2.0.0

# LangChain 相關 (使用相容版本)
langchain==0.2.16
langchain-google-genai==1.0.10
langchain-pinecone==0.1.3
langchain-community==0.2.16

# 向量數據庫
pinecone-client==4.1.1

# Google Generative AI
google-generativeai<0.8.0

# 嵌入模型
sentence-transformers>=2.2.0

# PDF 處理
PyPDF2>=3.0.0
pdfplumber>=0.9.0

# 其他工具
numpy<2.0.0
python-dotenv==1.0.1

# 額外的依賴
tiktoken==0.7.0
typing-extensions>=4.0.0

# 安全和檔案處理
Werkzeug>=2.0.0